Initial check-in of SSD1306 project.
This is the initial check-in of the SSD1306 project. It still needs documentation and some of the routines still need work, but it's complete enough to give a reasonable example of controlling cheap SSD1306 displays available off Amazon from a simple circuit board and the PIC32MX MPU.

This also includes a simple I2C two-wire implementation that uses interrupts, as well as a basic graphics display class that controls an SSD1306-based 128x64 display, along with the circuit board layout for the system.

Signed-off-by: William Woody <woody@alumni.caltech.edu>